ROMIO: construct the I/O aggregator rank list for Lustre driver.
This commit changes the algorithm of I/O aggregator selection. It first
requires the value of Lustre file stripe count, and set the number of
aggregators always equal to the stripe count, ignoring the MPI-IO hint
'cb_nodes' if set by users. The algorithm selects I/O aggregators from
different compute nodes in a round-robin fashion, one process at a time,
until the number of aggregators equal to the number of file stripe
count. The old method selects processes from a node first before moving
to the next node (i.e. OpenMP style). When the number of nodes is less
than the file stripe count, this new method still selects the number of
I/O aggregators same as the stripe count, versus the old method that
selects one process per compute node. When the number of nodes is
smaller than stripe count, the selected processes in each node are
spread evenly, rather than in continuous rank IDs. Similarly for when
the number of nodes is larger than the stripe count.