ROMIO: change the I/O aggregator rank list for Lustre driver.
This commit changes the algorithm of I/O aggregator selection to the
followings. It first acquires the Lustre file striping count, and then
uses it to set the number of aggregators always equal to the striping
count, ignoring the MPI-IO hint 'cb_nodes' if set by users. The
selection is to select processes on different compute nodes in a
round-robin fashion, one process at a time, until the number of
aggregators equal to the striping count. Note the old algorithm selects
processes from a node first before moving to the next node (i.e. OpenMP
style).  When the number of nodes is less than the file striping count,
this new algorithm will selects the number of I/O aggregators equal to
the striping count, versus the old method that selects one process per
compute node. When the number of nodes is smaller than striping count,
the new algorithm selects processes in each node spreading evenly (e.g.
0,4,8 will be selected on a node running 64 processes), rather than in
continuous rank IDs (e.g. 0,1,2 will be selected on a node running 64
processes). Similarly for when the number of nodes is larger than the
striping count, e.g. nodes 0,4,8 rather than 0,1,2.