Improve ADIOI_LUSTRE_Fill_send_buffer
ADIOI_LUSTRE_Fill_send_buffer can be expensive when the number of
noncontiguous segments to be copied from the user buffer to an
internally allocated send buffer is large.

Note even if the entire user buffer is noncontiguous, because a
collective write is performed in multiple rounds of 2-phase I/O,
in some rounds, the part of user buffer may be contiguous to be
used in MPI_Issend directly.

This commit avoids such memcpy.